Fixes to transaction balance handling in monthly DBs.
Modify logic of DB view credit_remaining and rename to sum_by_reason. It now returns group sums for each transaction type/reason. This primarily allows for detecting with certainty with 1 db query if there is a rollup. Assuming there is one when the sum is not negative is dangerous. New view also leaves opportunity for quering for other purposes like statistics.

kazoo_modb: Modify to try old view if the new view doesn't exist. Not worth rebuilding views since it should be a 1 time query on an old db.
get_modn to support for getting the month and year in either text or int format.

wht_util: 
*Remove the export of get_balance_from_account since it should only be an internal function that is only used once ever per account to get the balance from the accounts primary db incase of upgrade from pre modb transaction storage (pre 3.09).
* Redo rollup logic. There were two functions for doing rollups. Possible for a month db to have transactions in a month, there not be a rollup transaction and when the rollup was done the balance from a month without a rollup would be ignored/lost.
* Support old view or new view response.
* Only refresh views if needed for current modb.